Axel bit
Code of the axel rod projects code for the game that they made to work with the microbit and to be able to run this in full on both microbits. This will be downloadable to the micro bits throught the .hex file that there is or to go onto the micro bit editor (https://pxt.microbit.org/?lang=en).